music-1st Batch
Voicegroup 191 will be used for all songs from here on out, as I've tried to put every instrument into that one group. I probably missed some that I will have to put in later, but it covers most of the emerald instruments. None of the 3 new songs are on any of the maps right now, but they are in the game ready to be used. mus_anville is Anville Town from BW, the other two are original songs.